+++ src/main/java/entidades/ProdutoFisico.java
@@ -5,7 +5,7 @@ public class ProdutoFisico extends Produto{
     private Float altura;
     private Float largura;
 
-    public ProdutoFisico(Float preco, String nome, Integer codigoProduto, Float peso, Float altura, Float largura) {
+    public ProdutoFisico(Double preco, String nome, Integer codigoProduto, Float peso, Float altura, Float largura) {
         super(preco, nome, codigoProduto);
         this.altura = altura;
         this.peso = peso;
